Rethrow original error with attached extensions
This fixes our stacktraces, and a regression with our ability to set the `code` field
in the graphql error extensions. It currently only reports `INTERNAL_SERVER_ERROR`, which
is likely causing problems elsewhere.

This regression has been fixed AFAICT in graphql-tools#master, but occurs due to a series
of bugs in the `transformSchema` operation.

By not using `locatedError` and instead constructing a GraphQLError here, we can preserve
the original extensions and stacktraces, where they were previously being reset.